requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
idna==3.10
psycopg==3.2.9
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
